+++ pfnet/tests/utils.py
@@ -247,6 +247,8 @@ def compare_buses(test, bus1, bus2, check_internals=False, check_indices=True, e
     test.assertEqual(bus1.is_regulated_by_gen(),bus2.is_regulated_by_gen())
     test.assertEqual(bus1.is_regulated_by_tran(),bus2.is_regulated_by_tran())
     test.assertEqual(bus1.is_regulated_by_shunt(),bus2.is_regulated_by_shunt())
+    test.assertEqual(bus1.is_regulated_by_facts(),bus2.is_regulated_by_facts())
+    test.assertEqual(bus1.is_regulated_by_vsc_converter(),bus2.is_regulated_by_vsc_converter())
     test.assertLess(norminf(bus1.price-bus2.price), eps)
     test.assertEqual(len(bus1.generators),len(bus2.generators))
     test.assertEqual(len(bus1.reg_generators),len(bus2.reg_generators))
@@ -258,6 +260,11 @@ def compare_buses(test, bus1, bus2, check_internals=False, check_indices=True, e
     test.assertEqual(len(bus1.var_generators),len(bus2.var_generators))
     test.assertEqual(len(bus1.reg_trans),len(bus2.reg_trans))
     test.assertEqual(len(bus1.reg_shunts),len(bus2.reg_shunts))
+    test.assertEqual(len(bus1.csc_converters), len(bus2.csc_converters))
+    test.assertEqual(len(bus1.vsc_converters), len(bus2.vsc_converters))
+    test.assertEqual(len(bus1.reg_vsc_converters), len(bus2.reg_vsc_converters))
+    test.assertEqual(len(bus1.facts), len(bus2.facts))
+    test.assertEqual(len(bus1.reg_facts), len(bus2.reg_facts))
     if check_indices:
         test.assertEqual(set([o.index for o in bus1.generators]),
                          set([o.index for o in bus2.generators]))
@@ -279,6 +286,16 @@ def compare_buses(test, bus1, bus2, check_internals=False, check_indices=True, e
                          set([o.index for o in bus2.var_generators]))
         test.assertEqual(set([o.index for o in bus1.batteries]),
                          set([o.index for o in bus2.batteries]))
+        test.assertEqual(set([o.index for o in bus1.csc_converters]),
+                         set([o.index for o in bus2.csc_converters]))
+        test.assertEqual(set([o.index for o in bus1.vsc_converters]),
+                         set([o.index for o in bus2.vsc_converters]))
+        test.assertEqual(set([o.index for o in bus1.reg_vsc_converters]),
+                         set([o.index for o in bus2.reg_vsc_converters]))
+        test.assertEqual(set([o.index for o in bus1.facts]),
+                         set([o.index for o in bus2.facts]))
+        test.assertEqual(set([o.index for o in bus1.reg_facts]),
+                         set([o.index for o in bus2.reg_facts]))
     test.assertLess(norminf(bus1.sens_P_balance-bus2.sens_P_balance), eps)
     test.assertLess(norminf(bus1.sens_Q_balance-bus2.sens_Q_balance), eps)
     test.assertLess(norminf(bus1.sens_v_mag_u_bound-bus2.sens_v_mag_u_bound), eps)
@@ -305,7 +322,7 @@ def compare_generators(test, gen1, gen2, check_internals=False, eps=1e-10):
     test : unittest.TestCase
     gen1 : |Generator|
     gen2 : |Generator|
-    check_configuration : |TrueFalse|
+    check_internals : |TrueFalse|
     eps : float
     """
     
@@ -352,7 +369,7 @@ def compare_loads(test, load1, load2, check_internals=False, eps=1e-10):
     test : unittest.TestCase
     load1 : |Load|
     load2 : |Load|
-    check_configuration : |TrueFalse|
+    check_internals : |TrueFalse|
     eps : float
     """
 
@@ -398,7 +415,7 @@ def compare_shunts(test, shunt1, shunt2, check_internals=False, eps=1e-10):
     test : unittest.TestCase
     shunt1 : |Shunt|
     shunt2 : |Shunt|
-    check_configuration : |TrueFalse|
+    check_internals : |TrueFalse|
     eps : float
     """
 
@@ -432,7 +449,7 @@ def compare_branches(test, branch1, branch2, check_internals=False, eps=1e-10):
     test : unittest.TestCase
     branch1 : |Branch|
     branch2 : |Branch|
-    check_configuration : |TrueFalse|
+    check_internals : |TrueFalse|
     eps : float
     """
 
@@ -482,7 +499,250 @@ def compare_branches(test, branch1, branch2, check_internals=False, eps=1e-10):
         test.assertLess(norminf(branch1.sens_ratio_l_bound-branch2.sens_ratio_l_bound), eps)
         test.assertLess(norminf(branch1.sens_phase_u_bound-branch2.sens_phase_u_bound), eps)
         test.assertLess(norminf(branch1.sens_phase_l_bound-branch2.sens_phase_l_bound), eps)
+
+def compare_dc_buses(test, bus1, bus2, check_internals=False, check_indices=True, eps=1e-10):
+    """
+    Method for checking if two |BusDC| objects are similar.
+    
+    Parameters
+    ----------
+    test : unittest.TestCase
+    bus1 : |BusDC|
+    bus2 : |BusDC|
+    check_internals : |TrueFalse|
+    check_indices : |TrueFalse|
+    eps : float
+    """
+
+    test.assertTrue(bus1 is not bus2)
+    test.assertEqual(bus1.number, bus2.number)
+    test.assertEqual(bus1.num_periods, bus2.num_periods)
+    test.assertEqual(bus1.name.upper().strip(), bus2.name.upper().strip())
+    test.assertLess(norminf(bus1.v_base-bus2.v_base), eps)
+    test.assertLess(norminf(bus1.v-bus2.v), eps)
+    test.assertEqual(len(bus1.csc_converters),len(bus2.csc_converters))
+    test.assertEqual(len(bus1.vsc_converters),len(bus2.vsc_converters))
+    test.assertEqual(len(bus1.branches_k),len(bus2.branches_k))
+    test.assertEqual(len(bus1.branches_m),len(bus2.branches_m))
+    if check_indices:
+        test.assertEqual(set([o.index for o in bus1.csc_converters]),
+                         set([o.index for o in bus2.csc_converters]))
+        test.assertEqual(set([o.index for o in bus1.vsc_converters]),
+                         set([o.index for o in bus2.vsc_converters]))
+        test.assertEqual(set([o.index for o in bus1.branches_k]),
+                         set([o.index for o in bus2.branches_k]))
+        test.assertEqual(set([o.index for o in bus1.branches_m]),
+                         set([o.index for o in bus2.branches_m]))
+    if check_internals:
+        test.assertLess(norminf(bus1.index_v-bus2.index_v),eps)
+        test.assertEqual(bus1.flags_vars,bus2.flags_vars)
+        test.assertEqual(bus1.flags_fixed,bus2.flags_fixed)
+        test.assertEqual(bus1.flags_bounded,bus2.flags_bounded)
+        test.assertEqual(bus1.flags_sparse,bus2.flags_sparse)
+
+def compare_dc_branches(test, branch1, branch2, check_internals=False, eps=1e-10):
+    """
+    Method for checking if two |BranchDC| objects are similar.
+    
+    Parameters
+    ----------
+    test : unittest.TestCase
+    branch1 : |BranchDC|
+    branch2 : |BranchDC|
+    check_internals : |TrueFalse|
+    eps : float
+    """
+
+    test.assertTrue(branch1 is not branch2)
+    test.assertEqual(branch1.name, branch2.name)
+    test.assertEqual(branch1.num_periods, branch2.num_periods)
+    test.assertEqual(branch1.bus_k.index, branch2.bus_k.index)
+    test.assertEqual(branch1.bus_m.index, branch2.bus_m.index)
+    test.assertLess(norminf(branch1.r-branch2.r), eps*(1+norminf(branch1.r)))
+    if check_internals:
+        test.assertEqual(branch1.flags_vars,branch2.flags_vars)
+        test.assertEqual(branch1.flags_fixed,branch2.flags_fixed)
+        test.assertEqual(branch1.flags_bounded,branch2.flags_bounded)
+        test.assertEqual(branch1.flags_sparse,branch2.flags_sparse)
+
+def compare_csc_converters(test, conv1, conv2, check_internals=False, eps=1e-10):
+    """
+    Method for checking if two |ConverterCSC| objects are similar.
+    
+    Parameters
+    ----------
+    test : unittest.TestCase
+    conv1 : |ConverterCSC|
+    conv2 : |ConverterCSC|
+    check_internals : |TrueFalse|
+    eps : float
+    """
+
+    test.assertTrue(conv1 is not conv2)
+    test.assertEqual(conv1.name, conv2.name)
+    test.assertEqual(conv1.num_periods, conv2.num_periods)
+    test.assertEqual(conv1.ac_bus.index, conv2.ac_bus.index)
+    test.assertEqual(conv1.dc_bus.index, conv2.dc_bus.index)
+    test.assertLess(norminf(conv1.P_dc_set-conv2.P_dc_set), eps)
+    test.assertLess(norminf(conv1.i_dc_set-conv2.i_dc_set), eps)
+    test.assertLess(norminf(conv1.v_dc_set-conv2.v_dc_set), eps)
+    test.assertLess(norminf(conv1.P-conv2.P), eps)
+    test.assertLess(norminf(conv1.Q-conv2.Q), eps)
+    test.assertLess(norminf(conv1.P_dc-conv2.P_dc), eps)
+    test.assertLess(norminf(conv1.num_bridges-conv2.num_bridges), eps)
+    test.assertLess(norminf(conv1.x_cap-conv2.x_cap), eps)
+    test.assertLess(norminf(conv1.x-conv2.x), eps)
+    test.assertLess(norminf(conv1.r-conv2.r), eps)
+    test.assertLess(norminf(conv1.ratio-conv2.ratio), eps)
+    test.assertLess(norminf(conv1.ratio_max-conv2.ratio_max), eps)
+    test.assertLess(norminf(conv1.ratio_min-conv2.ratio_min), eps)
+    test.assertLess(norminf(conv1.angle-conv2.angle), eps)
+    test.assertLess(norminf(conv1.angle_max-conv2.angle_max), eps)
+    test.assertLess(norminf(conv1.angle_min-conv2.angle_min), eps)
+    test.assertLess(norminf(conv1.v_base_p-conv2.v_base_p), eps)
+    test.assertLess(norminf(conv1.v_base_s-conv2.v_base_s), eps)
+    
+    test.assertEqual(conv1.is_rectifier(), conv2.is_rectifier())
+    test.assertEqual(conv1.is_inverter(), conv2.is_inverter())
+    test.assertEqual(conv1.is_in_P_dc_mode(), conv2.is_in_P_dc_mode())
+    test.assertEqual(conv1.is_in_i_dc_mode(), conv2.is_in_i_dc_mode())
+    test.assertEqual(conv1.is_in_v_dc_mode(), conv2.is_in_v_dc_mode())
+
+    if check_internals:
+        test.assertLess(norminf(conv1.index_P-conv2.index_P),eps)
+        test.assertLess(norminf(conv1.index_Q-conv2.index_Q),eps)
+        test.assertLess(norminf(conv1.index_P_dc-conv2.index_P_dc),eps)
+        test.assertLess(norminf(conv1.index_i_dc-conv2.index_i_dc),eps)
+        test.assertLess(norminf(conv1.index_ratio-conv2.index_ratio),eps)
+        test.assertLess(norminf(conv1.index_angle-conv2.index_angle),eps)
+        test.assertEqual(conv1.flags_vars,conv2.flags_vars)
+        test.assertEqual(conv1.flags_fixed,conv2.flags_fixed)
+        test.assertEqual(conv1.flags_bounded,conv2.flags_bounded)
+        test.assertEqual(conv1.flags_sparse,conv2.flags_sparse)
+
+def compare_vsc_converters(test, conv1, conv2, check_internals=False, eps=1e-10):
+    """
+    Method for checking if two |ConverterVSC| objects are similar.
+    
+    Parameters
+    ----------
+    test : unittest.TestCase
+    conv1 : |ConverterVSC|
+    conv2 : |ConverterVSC|
+    check_internals : |TrueFalse|
+    eps : float
+    """
+
+    test.assertTrue(conv1 is not conv2)
+    test.assertEqual(conv1.name, conv2.name)
+    test.assertEqual(conv1.num_periods, conv2.num_periods)
+    test.assertEqual(conv1.ac_bus.index, conv2.ac_bus.index)
+    test.assertEqual(conv1.dc_bus.index, conv2.dc_bus.index)
+    if conv1.reg_bus is not None:
+        test.assertEqual(conv1.reg_bus.index,conv2.reg_bus.index)
+    else:
+        test.assertTrue(conv2.reg_bus is None)
+    test.assertLess(norminf(conv1.P_dc_set-conv2.P_dc_set), eps)
+    test.assertLess(norminf(conv1.v_dc_set-conv2.v_dc_set), eps)
+    test.assertLess(norminf(conv1.P-conv2.P), eps)
+    test.assertLess(norminf(conv1.Q-conv2.Q), eps)
+    test.assertLess(norminf(conv1.P_dc-conv2.P_dc), eps)
+    test.assertLess(norminf(conv1.loss_coeff_A-conv2.loss_coeff_A), eps)
+    test.assertLess(norminf(conv1.loss_coeff_B-conv2.loss_coeff_B), eps)
+    test.assertLess(norminf(conv1.P_max-conv2.P_max), eps)
+    test.assertLess(norminf(conv1.P_min-conv2.P_min), eps)
+    test.assertLess(norminf(conv1.Q_max-conv2.Q_max), eps)
+    test.assertLess(norminf(conv1.Q_min-conv2.Q_min), eps)
+    test.assertLess(norminf(conv1.Q_par-conv2.Q_par), eps)
+    test.assertLess(norminf(conv1.target_power_factor-conv2.target_power_factor), eps)
+
+    test.assertEqual(conv1.is_in_f_ac_mode(), conv2.is_in_f_ac_mode())
+    test.assertEqual(conv1.is_in_v_ac_mode(), conv2.is_in_v_ac_mode())
+    test.assertEqual(conv1.is_in_P_dc_mode(), conv2.is_in_P_dc_mode())
+    test.assertEqual(conv1.is_in_v_dc_mode(), conv2.is_in_v_dc_mode())
+
+    if check_internals:
+        test.assertLess(norminf(conv1.index_P-conv2.index_P),eps)
+        test.assertLess(norminf(conv1.index_Q-conv2.index_Q),eps)
+        test.assertLess(norminf(conv1.index_P_dc-conv2.index_P_dc),eps)
+        test.assertLess(norminf(conv1.index_i_dc-conv2.index_i_dc),eps)
+        test.assertEqual(conv1.flags_vars,conv2.flags_vars)
+        test.assertEqual(conv1.flags_fixed,conv2.flags_fixed)
+        test.assertEqual(conv1.flags_bounded,conv2.flags_bounded)
+        test.assertEqual(conv1.flags_sparse,conv2.flags_sparse)
+
+def compare_facts(test, facts1, facts2, check_internals=False, eps=1e-10):
+    """
+    Method for checking if two |Facts| objects are similar.
+    
+    Parameters
+    ----------
+    test : unittest.TestCase
+    facts1 : |Facts|
+    facts2 : |Facts|
+    check_internals : |TrueFalse|
+    eps : float
+    """
+
+    test.assertTrue(facts1 is not facts2)
+    
+    test.assertEqual(facts1.name, facts2.name)
+    test.assertEqual(facts1.num_periods, facts2.num_periods)
+    test.assertEqual(facts1.bus_k.index, facts2.bus_k.index)
+    if facts1.bus_m is not None:
+        test.assertEqual(facts1.bus_m.index, facts2.bus_m.index)
+    else:
+        test.assertTrue(facts2.bus_m is None)
+    if facts1.is_regulator():
+        test.assertEqual(facts1.reg_bus.index, facts2.reg_bus.index)
+    else:
+        test.assertTrue(facts2.reg_bus is None)
         
+    test.assertEqual(facts1.is_STATCOM(), facts2.is_STATCOM())
+    test.assertEqual(facts1.is_SSSC(), facts2.is_SSSC())
+    test.assertEqual(facts1.is_UPFC(), facts2.is_UPFC())
+    test.assertEqual(facts1.is_series_link_disabled(), facts2.is_series_link_disabled())
+    test.assertEqual(facts1.is_series_link_bypassed(), facts2.is_series_link_bypassed())
+    test.assertEqual(facts1.is_in_normal_series_mode(), facts2.is_in_normal_series_mode())
+    test.assertEqual(facts1.is_in_constant_series_z_mode(), facts2.is_in_constant_series_z_mode())
+    test.assertEqual(facts1.is_in_constant_series_v_mode(), facts2.is_in_constant_series_v_mode())
+
+    test.assertLess(norminf(facts1.P_k-facts2.P_k), eps)
+    test.assertLess(norminf(facts1.Q_k-facts2.Q_k), eps)
+    test.assertLess(norminf(facts1.P_m-facts2.P_m), eps)
+    test.assertLess(norminf(facts1.Q_m-facts2.Q_m), eps)
+    test.assertLess(norminf(facts1.Q_sh-facts2.Q_sh), eps)
+    test.assertLess(norminf(facts1.Q_s-facts2.Q_s), eps)
+    test.assertLess(norminf(facts1.P_dc-facts2.P_dc), eps)
+    test.assertLess(norminf(facts1.Q_par-facts2.Q_par), eps)
+    test.assertLess(norminf(facts1.P_set-facts2.P_set), eps)
+    test.assertLess(norminf(facts1.Q_set-facts2.Q_set), eps)
+
+    test.assertLess(norminf(facts1.Q_max_s-facts2.Q_max_s), eps)
+    test.assertLess(norminf(facts1.Q_min_s-facts2.Q_min_s), eps)
+    test.assertLess(norminf(facts1.Q_max_sh-facts2.Q_max_sh), eps)
+    test.assertLess(norminf(facts1.Q_min_sh-facts2.Q_min_sh), eps)
+    test.assertLess(norminf(facts1.i_max_s-facts2.i_max_s), eps)
+    test.assertLess(norminf(facts1.i_max_sh-facts2.i_max_sh), eps)
+    test.assertLess(norminf(facts1.P_max_dc-facts2.P_max_dc), eps)
+    test.assertLess(norminf(facts1.v_min_m-facts2.v_min_m), eps)
+    test.assertLess(norminf(facts1.v_max_m-facts2.v_max_m), eps)
+
+    test.assertLess(norminf(facts1.v_mag_s-facts2.v_mag_s), eps)
+    test.assertLess(norminf(facts1.v_ang_s-facts2.v_ang_s), eps)
+    test.assertLess(norminf(facts1.v_max_s-facts2.v_max_s), eps)
+    test.assertLess(norminf(facts1.g-facts2.g), eps)
+    test.assertLess(norminf(facts1.b-facts2.b), eps)
+
+    if check_internals:
+        test.assertLess(norminf(facts1.index_P_k-facts2.index_P_k),eps)
+        test.assertLess(norminf(facts1.index_Q_k-facts2.index_Q_k),eps)
+        test.assertLess(norminf(facts1.index_P_m-facts2.index_P_m),eps)
+        test.assertLess(norminf(facts1.index_Q_m-facts2.index_Q_m),eps)
+        test.assertLess(norminf(facts1.index_Q_sh-facts2.index_Q_sh),eps)
+        test.assertLess(norminf(facts1.index_Q_s-facts2.index_Q_s),eps)
+        test.assertLess(norminf(facts1.index_P_dc-facts2.index_P_dc),eps)
+    
 def compare_networks(test, net1, net2, check_internals=False):
     """
     Method for checking if two |Network| objects are held in different
@@ -493,7 +753,7 @@ def compare_networks(test, net1, net2, check_internals=False):
     test : unittest.TestCase
     net1 : |Network|
     net2 : |Network|
-    check_configuration : |TrueFalse|
+    check_internals : |TrueFalse|
     """
 
     eps = 1e-10
@@ -514,29 +774,29 @@ def compare_networks(test, net1, net2, check_internals=False):
     # Buses
     test.assertEqual(net1.num_buses, net2.num_buses)
     for i in range(net1.num_buses):
-        bus1 = net1.buses[i]
-        bus2 = net2.buses[i]
+        bus1 = net1.get_bus(i)
+        bus2 = net2.get_bus(i)
         compare_buses(test, bus1, bus2, check_internals=check_internals, eps=eps) 
 
     # Branches
     test.assertEqual(net1.num_branches, net2.num_branches)
     for i in range(net1.num_branches):
-        branch1 = net1.branches[i]
-        branch2 = net2.branches[i]
+        branch1 = net1.get_branch(i)
+        branch2 = net2.get_branch(i)
         compare_branches(test, branch1, branch2, check_internals=check_internals, eps=eps)
 
     # Generators
     test.assertEqual(net1.num_generators, net2.num_generators)
     for i in range(net1.num_generators):
-        gen1 = net1.generators[i]
-        gen2 = net2.generators[i]
+        gen1 = net1.get_generator(i)
+        gen2 = net2.get_generator(i)
         compare_generators(test, gen1, gen2, check_internals=check_internals, eps=eps)
         
     # Var generators
     test.assertEqual(net1.num_var_generators, net2.num_var_generators)
     for i in range(net1.num_var_generators):
-        vargen1 = net1.var_generators[i]
-        vargen2 = net2.var_generators[i]
+        vargen1 = net1.get_var_generator(i)
+        vargen2 = net2.get_var_generator(i)
         test.assertTrue(vargen1 is not vargen2)
         test.assertEqual(vargen1.num_periods, vargen2.num_periods)
         test.assertEqual(vargen1.bus.index, vargen2.bus.index)
@@ -560,22 +820,22 @@ def compare_networks(test, net1, net2, check_internals=False):
     # Shunts
     test.assertEqual(net1.num_shunts, net2.num_shunts)
     for i in range(net1.num_shunts):
-        shunt1 = net1.shunts[i]
-        shunt2 = net2.shunts[i]
+        shunt1 = net1.get_shunt(i)
+        shunt2 = net2.get_shunt(i)
         compare_shunts(test, shunt1, shunt2, check_internals=check_internals, eps=eps) 
 
     # Loads
     test.assertEqual(net1.num_loads, net2.num_loads)
     for i in range(net1.num_loads):
-        load1 = net1.loads[i]
-        load2 = net2.loads[i]
+        load1 = net1.get_load(i)
+        load2 = net2.get_load(i)
         compare_loads(test, load1, load2, check_internals=check_internals, eps=eps)
         
     # Batteries
     test.assertEqual(net1.num_batteries, net2.num_batteries)
     for i in range(net1.num_batteries):
-        bat1 = net1.batteries[i]
-        bat2 = net2.batteries[i]
+        bat1 = net1.get_battery(i)
+        bat2 = net2.get_battery(i)
         test.assertTrue(bat1 is not bat2)
         test.assertEqual(bat1.name, bat2.name)
         test.assertEqual(bat1.num_periods, bat2.num_periods)
@@ -598,10 +858,54 @@ def compare_networks(test, net1, net2, check_internals=False):
             test.assertEqual(bat1.flags_bounded,bat2.flags_bounded)
             test.assertEqual(bat1.flags_sparse,bat2.flags_sparse)
 
-    # Hashes                 
+    # DC Buses
+    test.assertEqual(net1.num_dc_buses, net2.num_dc_buses)
+    for i in range(net1.num_dc_buses):
+        bus1 = net1.get_dc_bus(i)
+        bus2 = net2.get_dc_bus(i)
+        compare_dc_buses(test, bus1, bus2, check_internals=check_internals, eps=eps)
+
+    # DC Branches
+    test.assertEqual(net1.num_dc_branches, net2.num_dc_branches)
+    for i in range(net1.num_dc_branches):
+        br1 = net1.get_dc_branch(i)
+        br2 = net2.get_dc_branch(i)
+        compare_dc_branches(test, br1, br2, check_internals=check_internals, eps=eps)
+
+    # CSC converters
+    test.assertEqual(net1.num_csc_converters, net2.num_csc_converters)
+    for i in range(net1.num_csc_converters):
+        conv1 = net1.get_csc_converter(i)
+        conv2 = net2.get_csc_converter(i)
+        compare_csc_converters(test, conv1, conv2, check_internals=check_internals, eps=eps)
+
+    # VSC converters
+    test.assertEqual(net1.num_vsc_converters, net2.num_vsc_converters)
+    for i in range(net1.num_vsc_converters):
+        conv1 = net1.get_vsc_converter(i)
+        conv2 = net2.get_vsc_converter(i)
+        compare_vsc_converters(test, conv1, conv2, check_internals=check_internals, eps=eps)
+        
+    # FACTS
+    test.assertEqual(net1.num_facts, net2.num_facts)
+    for i in range(net1.num_facts):
+        facts1 = net1.get_facts(i)
+        facts2 = net2.get_facts(i)
+        compare_facts(test, facts1, facts2, check_internals=check_internals, eps=eps)
+        
+    # Hashes (AC)                
     for bus in net1.buses:
+        test.assertEqual(bus.number,net1.get_bus_from_number(bus.number).number)
+        test.assertEqual(bus.name,net1.get_bus_from_number(bus.number).name)
         test.assertEqual(bus.number,net2.get_bus_from_number(bus.number).number)
         test.assertEqual(bus.name,net2.get_bus_from_number(bus.number).name)
+
+    # Hashes (DC)
+    for bus in net1.dc_buses:
+        test.assertEqual(bus.number,net1.get_dc_bus_from_number(bus.number).number)
+        test.assertEqual(bus.name,net1.get_dc_bus_from_number(bus.number).name)
+        test.assertEqual(bus.number,net2.get_dc_bus_from_number(bus.number).number)
+        test.assertEqual(bus.name,net2.get_dc_bus_from_number(bus.number).name)
 
 def check_network(test, net):
     """