+++ CHANGELOG.rst
@@ -26,6 +26,7 @@ Unreleased
 * Added "is_in_service" method and "in_service" attribute to load.
 * Exposed routines for updated load P and Q components according to provided weights.    
 * Exposed CSC HVDC constraints/functions.
+* Extended test utils to compare dc buses, dc branches, FACTS, csc converters, and vsc converters.    
   
 Version 1.3.3
 -------------